+++ test/specs/components/Indicator/Indicator-test.ts
@@ -1,0 +1,7 @@
+import { isConformant } from 'test/specs/commonTests'
+
+import Indicator from 'src/components/Indicator/Indicator'
+
+describe('Indicator', () => {
+  isConformant(Indicator)
+})